refactor(view): 並び替え処理をhooksに切り出し
